move vpsdb to it's own lib for code check reuse
